RavenDB-10799 rename and make class private
